+++ components/AzTable.vue
@@ -11,38 +11,41 @@
   />
 </template>
 <script setup lang="ts">
-import { Table } from "ant-design-vue";
-import useAxios from "../services/axios";
-import { useOffsetPagination } from "@vueuse/core";
-import { ref, onMounted, reactive } from "vue";
+import { Table } from 'ant-design-vue'
+import useAxios from '../services/axios'
+import { useOffsetPagination } from '@vueuse/core'
+import { ref, onMounted, reactive } from 'vue'
+import type { ColumnsType, TablePaginationConfig } from 'ant-design-vue/es/table'
 interface Props {
-  url: string;
+  url: string
+  columns: ColumnsType
 }
 
 const props = withDefaults(defineProps<Props>(), {
-  url: "",
-});
+  url: '',
+  columns: () => []
+})
 
-const axios = useAxios();
-const data = ref([]);
+const axios = useAxios()
+const data = ref([])
 const fetchData = async (page: number, pageSize: number) => {
   try {
     const result = await axios.post({
       url: props.url,
       data: {
         size: pageSize,
-        page: page,
-      },
-    });
+        page: page
+      }
+    })
     data.value = result.data.map((item: any) => ({
       ...item,
-      key: item.id,
-    }));
-    pagination.value.total = result.content.length;
+      key: item.id
+    }))
+    pagination.total = result.content.length
   } catch (error) {
-    console.error("Error fetching data:", error);
+    console.error('Error fetching data:', error)
   }
-};
+}
 
 const pagination = reactive({
   showSizeChanger: true,
@@ -51,32 +54,34 @@ const pagination = reactive({
   pageSizeOptions: [5, 10, 20, 30, 50],
   current: 1,
   pageSize: 5,
-  total: 0,
-});
+  total: 0
+})
 
 const { currentPage, currentPageSize } = useOffsetPagination({
   total: 0,
   page: 1,
   pageSize: 5,
-  onPageChange: (page) => {
-    pagination.current = page;
-    fetchData(page, currentPageSize.value);
+  onPageChange: (_data) => {
+    pagination.current = _data.currentPage
+    pagination.pageSize = _data.currentPageSize
+    fetchData(_data.currentPage, _data.currentPageSize)
   },
-  onPageSizeChange: (pageSize) => {
-    pagination.pageSize = pageSize;
-    fetchData(currentPage.value, pageSize);
-  },
-});
+  onPageSizeChange: (_data) => {
+    pagination.current = _data.currentPage
+    pagination.pageSize = _data.currentPageSize
+    fetchData(_data.currentPage, _data.currentPageSize)
+  }
+})
 
 onMounted(() => {
-  fetchData(currentPage.value, currentPageSize.value);
-});
-const loading = ref(false);
-const handleTableChange = (pag, filters, sorter) => {
-  pagination.current = pag.current;
-  pagination.pageSize = pag.pageSize;
-  loading.value = true;
-  fetchData(pag.current, pag.pageSize);
-  loading.value = false;
-};
+  fetchData(currentPage.value, currentPageSize.value)
+})
+const loading = ref(false)
+const handleTableChange = (pag: TablePaginationConfig) => {
+  pagination.current = pag.current ?? 1
+  pagination.pageSize = pag.pageSize ?? 1
+  loading.value = true
+  fetchData(pag.current ?? 1, pag.pageSize ?? 1)
+  loading.value = false
+}
 </script>